Another Coverity "fix"...
This is kinda valid... I am reading a 32-bit integer from a socket,
then trying to loop that many times.  What Coverity didn't notice
though is that the loop will exit way before due to the buffer
being exhausted.

By ensuring that it's less than UINT32_MAX/4, I'm still not really
enforcing anything since with the 32-bit value for extcnt, and a
32-bit value for every string, there could only be strings before
it failed anyway, even with a 4GB buffer (which you couldn't do).