When editing the current/online user, copy the auto-term settings to user rec
(user_t.misc, User.settings), when enabling auto-terminal detection. This is
similar to what we do in logon.cpp already:

	useron.misc |= (AUTOTERM | autoterm);

We used to have this logic in useredit.cpp, but was removed in commit
8a71ab16c2, probably due to editing offline users (e.g. in ;uedit sysop
comand). So just do the copy of autoterm flags when editing the *current*
(online) user. This likely addresses the original concern in that commit.

And it fixes issue #853, thanks to Deuce for root-causing