Tie all the input and output threads together with an atomic_bool
Since we're using it in one place, use it everywhere.  While we're
here, read()/recv() of 0 after select indicates a closed connection
as well, an ensure we check terminate flags in inner loops.

This should help with SyncTERM issues where it appears to be
connected until you press a key, or it "hangs" and you have to
manually disconnect due to the connection actually having been
torn down at one end or the other.