Fix TLS short send issue.
On JS TLS sockets, sends over 16384 bytes would be truncated to
the next multiple of 8192 higher than half the buffer length.

This was triggered because we send chunks of 8192 bytes at a time,
and decrement the length each time through the loop.  We return
"success" when the total sent so far is higher than the length
remaining.

Fixes bug reported in #Synchronet by Accession.