Switch from socket_readable() to socket_check() in input threads
This allows us to detect the remote sending a FIN early and
disconnect rather than waiting until we have data to send.

This isn't actually necessary for the ones that use the socket
directly as they'll read zero bytes and terminate, but this way
it's consistent across all of them.

Another attempt to drive a stake through the heart of ticket 174