Add FileURL text.dat string, used when displaying file info
This is an optional (default: blank) string that can be used to display a
file's URL(s) (e.g. FTP and/or Web URLs, using @-codes) when viewing extended
file information. I currently have this string output before the file
description.

The idea is that a sysop could set this string (e.g in their ctrl/text.ini)
instead of or in-addition-to their text/menu/download.msg file and then the
file URL(s) would be more readily displayed (not just went prompting for a
download protocol to send said file).