Don't explicitly omit the frame pointer.
At -O3, the only reason the frame pointer will be retained is when
removing it makes it impossible to get any reasonable information
out of a core dump.

I personally prefer to take the small performance hit on affected
platforms and always include the frame pointer because it makes
profiling release builds possible... but I can manually make that
adjustment when I want to.