Add "native" toggle option to Viewable/Extractable/Testable/Compressible files
... and download events.

The "Native Program List" was a hack left over from the transition from
MS-DOS to OS/2 and now with extensible configuration files, like we have now,
there's no reason not to add more fields and make it easier to specify which
of these programs (most/all of them nowadays) are actually native to the
host OS.

The defaults (e.g. when creating a new file type) are unchanged for now.