Another large speedup... only update bitmap when drawing a frame
Previously, every change would be flushed through to the bitmap, now
it only flushes before a frame is sent to the driver, or when specific
things (like setting pixels) occur.

This about doubles the throughput again, running the test at 11
seconds (started at 179).  At a 16× speedup now from when I started.

This can now display at about 10Mbps, which is still kinda slow,
but at least it's not completely terrible anymore.

This could likely be improved by better timing of the frame
generation... it tries to hot 100Hz right now, which is a bit
excessive.