Make vstat vmem a ringbuffer as well.
This shaves off another .75 seconds, and gets rid of the last
glaring plateau in the stack flamegraph.

What's interersting is that I'm seeing my CPU utilization drop now
too, which implies I've reached the speed of the pty connection
type, or the speed of the term worker loop.  Maybe I'll look there
next, but since the pty isn't a major use-case for SyncTERM, I may
not... at this point it's just benchmark golf.