Reduce audio level by -12dB.
It was at absolute max (an extra one and it would clip) which is
really a terrible thing to do to people wearing headphones.

I don't really need to do it this way, but the math for the different
wave forms depends heavily on integer wrapping (and WAVE_SHAPE_SINE_HARM
is just stupid).