Add new "Binmode Broken" option for telnet.
In this mode, SyncTERM won't will/do TELNET_BINARY_TX at the start
of a connection, leaving the mode up to the remote system. The main
purpose for this option is to work around a bug in older Synchronet
releases where early CTRL-C checking wouldn't work in binary mode.

Implements feature request 91.