Throw fewer exceptions when passed null or undefined as parameters
Some scripts (e.g. addfiles.js) depend on the previous (undocumented) behavior
for some methods, e.g. lfexpand(undefined) would return undefined.

So for global functions that are expected to return a modified version of the
value passed (usually a string), return null or undefined when passed null or
undefined. Backward compatibility is more important than consistency in this
case. :-(