Add per-service option: NO_USER_PROT
If enabled, this option prevents logins to this service from changing the
protocol (aka modem) field for a user logging-in.

This is a work around for Deon who has some message areas set to require
"PROT NOT IMAP" for access requirements and when his IMAP client connects
and logs-in, that changes the protocol field in his user account (to "IMAP")
which means that user account cannot see/read those message areas (until the
protocol field is changed by another login to another server/service) and
thus won't receive echomail notifications for message received for them in
those areas.