Fix up cursor and blinking rates, step 1...
For "PC" modes, cursor blinked at 1/16th of the VSYNC rate, so use
the CGA timings, where were generally the slowest of the bunch...
This means cursor blink at 3.745Hz, and character blink at 1.8725Hz.

For Prestel, character blink was 0.75Hz with a 3:1 On:Off ratio.
The cursor on the other hand was 1/32th the VSYNC and was generally
PAL, so 1.5625Hz.

I still need to dig into C64 and Atari modes, because they're likely
way off now (assuming either supports blinking).