Add/use new function: listAddNodeWithFlags()
Fixes race condition/possible crash in listAddNodeData(), listAddNodeString()
and listAddNodeList(). These functions were modifying the node->flags *after*
a node was added to the list and list lock released.

This is most likely the cause of the issue caught by valgrind when running the
jsexec-testsuite:
https://gitlab.synchro.net/main/sbbs/-/jobs/498286

Thank you Deuce and valgrind, nice catch!