Add new "PostTo" setting to (optionally) change the recipient of SMTP-posts
Messages posted to sub-boards via SMTP normally will have the recipient
name/address as the "To" (RECIPIENT) value of the posted message. If the sysop
wishes to replace this string (for all SMTP-posted messages) with a different
string (e.g. "All") this is the setting to allow that.

I do want web-scrapping spammers to collect the 'to' fields for my TLDR and
Coverity-scan posts and post their own spam messages via email (any more).

Also accommodate the new link list API (long -> int).