Attempt to silence false positive warning.
The warning suggests that x1 and y1 may be uninitialized in the i > 0
block, but that's not really possible...

It's too bad the warning doesn't clarify.  It's also weird that the
warning was in the draw_line() call and not in the lines above that
(incorrectly) compared them with -1.

Fix that check while we're here.