Add new "BBC Micro" screen mode.
This is almost exactly like Prestel, but it's for BBS Micro BBSs
such as BeeBS.  The basic changes are:
1) Moving off the bottom of the screen scrolls instead of wraps
2) BBC Micro "high ASCII" is supported
3) Return sends \r, not #

To support the high ASCII, I actually implemented the ITU-T Rec.T0101
G1 set, which means the smooth mosaic characters are theoretically
available.

However, the SAA5050 emulation clobbers this by "correctly" turning
them into "blast-through" alphanumeric characters.

Doing this required the separated mosaic support be moved into the
font rendering and out of the font itself.

Anyway, the Sysop of beebs.ddns.net should be happy... and since
this will advertise a terminal type of "BBC Micro", they should be
able to auto-detect between ANSI and BBC Micro during login so the
user doesn't need to enter it.  Hit me up on IRC if you want help
with that.