Split the happy path font rendering into separate function
And merge the single/double height ones into the same function.
Normalize all the types while I'm here.

With this, the happy path draw_char_row_fast() is very simple and
easy to keep updated, and all of the complexity goes into
draw_char_row_slow().

There seems to be around a 20% performance difference between them,
but it can be hard to tell for sure because of the time spent
cheating.

Note that it doesn't appear to be worth cheating if we're going fast,
but that's getting a bit too deep for my tastes.