Allow "reserving" names
The SyncTERM cache uses the BBS name as the directory name, so any
system-level caching needs to either be in the cache root directory
or in a directory whose name is guaranteed to not collide with an
entry name.

This leaves two options... either prevent a list of suffixes from
being used on BBS names, or have a separate system cache whose name
cannot be used as a BBS name.

I've taken the second route here... SyncTERM will not load a BBS
entry with the (case insensitive) name "syncterm-system-cache" and
will not allow creating one.  This is the directory where "global"
cache files will be stored (ie: scripts, internet dialing
directories, etc.)