Call exit() rather than sleep forever in retro_deinit()
This doesn't fix anything since calling exit() from a DLL is
nonsense, but at least it's less wrong than just waiting forever.

Also, when main() returns, tell the frontent to shut down.

It's most likely that this ends up causing the frontend to crash,
but if the stars align, it may exit instead.