Better duplicate "key value" detection/rejection for configuration settings
Internal codes are always treated as keys (in a database sense) and should be
unique, but they weren't always enforced to be unique in SCFG. There's still
a few config items with internal codes that aren't forced unique (e.g. chat
channels).

File Library short names and Mesage Group short names are also key values and
must be unique and that's now enforced as well.

Remove the default name for new File Libraries and Message Groups (which are
likely duplicate anyway).