Fix bug with fixed event toggles
I'm totally perplexed that none of the compilers we have warned about this
obvious incorrect usage of the toggle_flag() function. Not warning about
passing false (0/NULL) for a char* makes some sense, but not the passing of
a const char* in place of a bool (int)? Maybe we should be building with
-Wall?