Correctly center lines of text that contain @-codes, after @center@
@-codes weren't expanded before the text length calculation, so the centering
logic (offset in the user's terminal screen) would usually be incorrect as
pointed out by Nelgin in #synchronet when we were discussing the |C @-code
modifier (which is to center an @-code value in a field, not the screen).

This issue was a known limitation of the fix to issue #418 (commit 8987150bd)
but now that we have the expand_atcodes() function (hooray!), this is an
easily solvable problem.

Note: we're now calling center() withOUT the 'msg' parameter set to true,
which means the @center@ code goes back to the previous (to issue #418 fix)
behavior of just using bputs() for the centered text (we don't need the
recursive call to pustmsg() to support @-codes any longer). This was the only
reason that center() grew the optional 'msg' parameter so we likely can remove
that option/feature at some date.