Target Windows Vista+ (not Windows XP) and enable enhanced instruction sets
Synchronet v3.20 doesn't actually work on Windows XP due to depenendies in
3rd party libraries we link with and more recent ciolib improvements.

So, no need to use an ancient WinSDK (7.0) or toolset anymore. Hooray!
I'm not changing the versions of C or C++ language enabled (yet), but that's
something we should consider doing.

Also, re-enable enhanced x86 processor extensions (e.g. SSE2) as I doubt very
much you can run Windows Vista on a pre-2000 CPU anyway. If this is a problem
for anyone, it's trivial to disable the processor extensions again.