When building for Windows Vista+, use GetTickCount64() through-out
At one point while migrating off the Windows XP compatible WinSDK and toolset,
I was getting nice deprecation warnings in this file about uses of
GetTickCount() and the 49 day roll-over isuse. I stopped getting those
deprecation warnings (not sure when/why), but I'm still doing the right thing
here/now and using the newer Win32 API function when it's available (always,
for the versions of Windows we're building Synchronet and friends for).