Support [modname:charset=<charset>] sections
This is a getting a little kludgey

I think doing something with ARS would be nicer, but then we get into the
area of constructing a 'options' object out of keys from multiple sections
and which sections should supercede which? This'll do for now for Nelgin's
BBC mode7 terminal project.