Fix Mode 7 (ie: SAA5050) mode double-height handling
Basically, the only thing the SAA5050 enforces is that characters
on a bottom row need to be double-height, or they're displayed as
spaces.

The Prestel terminal now has an extra bit that indicates it is a
prestel terminal, and the bitmap console hacks up the second row.

While we're poking this, use macros for magic bits in FG/BG colours.