Ignore MKD/XMKD commands from users (pretend successful)
For non-Guest users (without the 'U' restriction), if the FTP client
sends the MKD or XMKD command, just return the proper succcess result
and do nothing (don't log as a hack attempt or return a syntax error
result).

Some FTP clients (e.g. Reolink Doorbell camera) just always try to
blindly create a directory before uploading files, but if that command
fails (as we would), the upload is aborted. So pretend to create the
diretory whether it already exists or not, just so the client will
proceed with the upload.

For G or U restricted users, treat (log) as hack-attempt, as before.