Checkpoint
This may be the final interface...

A bit more scope creep, but it's sorting out now.  Once the ANSI
terminal class is done, it'll be possible to do a factory and a
transformer, which will allow plugging in to anywhere user.misc
is changed (which will likely suck).

Still worried about how invasive this will be in the face of a
dynamic user.misc.

OTOH, a RIP terminal class will be very interesting... and this
should pave the way to scrolling regions in the console object.