Use inet_addrtop to convert socket addresses to strings to handle endianness
Socket addresses are stored in network byte order (big endian) and this
sprintf() logic here assumed little endian.

This was a regression introduced in either commit 29b93c32 or
commit db8bb221 (in the past 2 weeks) when we stopped using (some)
of the deprecated IP address/resolution functions and fixed the socket address
storage to always be in network byte order (as is the norm).

This should fix the issue recently reported in IRC by BrokenMind.