Add CTRL-S to "save"
This simply writes changed values to ctrl/textedit.ini, you can
copy these into the appropriate text.ini or text.dat.

I haven't dug in to see if text.dat has different encoding than ini
or anything, and I think that if you want to use this to build a
full text.dat, you would want to do it a bit at a time, so a follow-
on utility to convert text.ini to text.dat is likely the easiest way
to do this stuff.