strtoattr() now ignores Ctrl-A chars in the string... as attrstr() used to
This returns bug-compatibility with attrstr(), at least for Ctrl-A chars,
so that if someone is for example, assigning a Ctrl-A sequence to JS
console.attributes, that'll work (again) as it did in SBBS v3.19.

Any other invalid chars in the string will stop the parsing, still, even
though that's not how attrstr() worked.

For Kirkman's doorgame/script backward compatibility.