When the player is not created, number should be zero
Should make issue #392 go away, but the underlying CNW bug which
will cause the player that runs maint to see all NPCs remains.

For CNW, maintenance should absolutely be ran as a timed event,
with a real name that never gets an account.  Luckily, if there's
no dropfile, real name defaults to SYSOP, which should meet that
criteria.