Set node "interrupt" flag to try to gracefully disconnect user blocking event
When a timed event is configured to run "exclusively", all nodes need to not be
in in-use. As it was, after waiting 60 minutes for the online user(s) to
notice they'd run out of time and disconnect, we'd just (rather ungracefully)
close the sockets used by such node(s) connections. This results in same logged
errors about trying to send to bad socket descriptors and provides no feedback
to the user about why they were disconnected.

Since we have the node interrupt flag (which hopefully, all scripts are
checking via node_sync) - use that to try to more gracefully terminate the
user's session/connection after 30 minutes of waiting for the user to
disconnect.

If after 60 minutes of waiting, the node is still in-use, we still do the
socket disconnection method.