Add node interrupt flag checking to *nix external() execution of stdio progs
and abort the execution of the program when the node is interrupted. This
should allow more cases were the node interrupt flag can be used to gracefully
disconnect a user running a DOS or native (stdio) program on *nix.

We have this functionality already in the Windows version of external()