More Rlogin xtrn-menu options (in modopts.ini [logon] section):
rlogin_xtrn_logon set to false to skip the normal logon screens for Rlogin.
rlogin_xtrn_logoff set to false (default) to immediately hangup, true to
prompt the user if they truly wish to logoff, and "full" to prompt and if
affirmed, go through the full logoff procedure (dislpay files and events,
etc.).