Fix SDL/GDI issue with textmode() using alternate fonts
Previously, SDL and GDI modes (but not X11 mode) would return early
from textmode() if the new mode is the same as the old mode. This
would result in the font not being reset to the defaults for the
mode.

This was most noticable when using RIP to change the font to one of
a difference size since the mode remains the same, but the font and
screen dimensions change.

Should fix issue 201