Fix default mode value (should *not* be 10, i.e TG_NODESYNC|TG_CRLF)
Bug introduced in commit 49053f3158e5a0b671c

the 'mode' value was by default, undefined.
the 'timeout' value is by default, 10.

When mode value/flags was not provided on the command-line, undefined
was passed to bbs.rlogin_gate() as the 5th parameter, but the number 10
is passed as the 6th parameter (for time-out). The problem is, the first
Number parameter passed to bbs.rlogin_gate() is interpretted as the mode
value and so that becomes 10 (0x0A) which includes TG_NODESYNC thus enabling
all node messages/activity being displayed to the rlogin user and interrupting
their rlogin session (e.g. game play).

Just make the 0 the default value for mode, like we did in telgate.js.