Store/reuse each server's client highwater mark for MQTT publishing
Although we always republished "0" as the client highwater mark (e.g. upon
server startup and recycle), the internally tracked highwater marked is *not*
reset (to zero) upon recycle, so may not have been republished (with a
non-zero) value until/unless the internal highwater mark was exceeded.

So now, the value we publish upon recycle will be same value as the internally
tracked client highwater mark.

This fixes issue #910