Fix screen save/restore for manual uploads
While it was properly entering the mode for the file picker, it
restored the screen before the protocol picker, which resulted in
the palette/charset from the connection being used for the transfer
window.

Fixes ticket 202

While we're here, wrap the whole "put everything in CP437 mode"
into a new function and use it everywhere.