Support multicolumn xtrn program *section* listing
While we've long supported multicolumn program lists (for long lists
that would likely cause a user's terminal to scroll), we've never supported
multicolumn external program *section* listings.

So if the multicolumn option is set to true (the default) and the list of
sections is longer than would fit on the user's terminal screen, split into
2 columns (just two).

This introduces the new configurable string-option: multicolumn_section_fmt

This fixes issue #915

Hopefully Nelgin will be in a better mood as a result.