Add Ctrl-A output encoding (now the default)
The output is CP437/Ctrl-A by default, use -u to force UTF-8 encoded characters
if you need that or -a to force ANSI output if you need that.

Removed the mIRC color support (that can be added back easiliy enough if anyone
actually wants that).

So the tdfiglet -e u|a option has been replaced with -u (UTF-8)
and -c m|a option has been replaced -a (ANSI). Simpler usage with correct
terminology.