If no font object or filename is passed to lib, pick a font file at random
This moves the random file selection into tdfonts_lib.js

We should also pick a font index at random (when opt.random is true), but
that's not happening yet.

Also fixed up some indentation issues (when using 4 space tabstops). The
original commit of this file used spaces for indentation, so it's kind of
a mess now and should be uncrustified.