Insure messages created/edited are always CRLF line-terminated
Even if a message editor is configured to *not* expand LF to CRLF (in SCFG),
when creating or editing message text, always do the LF->CRLF expansion, for
uniformity.

Allow the wmode flag (WM_*) to be passed to JS console.editfile() for use in
other scenarios where we want to insure the file is CRLF-terminated.