Support left/right justification
Part of issue/request #922

I don't know if we really want separate cycling/selection of foreground
and background colors (most foreground colors aren't good combos for
most background colors) or the separate cycling/selection of the upper,
mid, and lower parts of the border. These are doable, but I'm not sure
the majority of the results would look good.