Add support to SMBLIB for "Fixed" format message text
By default, messages are assumed to be of "Flowed" format (CRLF-delimited
paragraphs). When posting pre-formatted text, this aux attribute flag
should be set to indicate to message viewers that the message text should
be displayed as-is, without applying word-wrapping or parsing/expanding
markup tags.

This adds a new key/value field to QWK HEADERS.DAT and FTN Kludge:
"Format: fixed" and "Fixed: flowed" as appropriate.

This flag is now automatically set when posting messages in raw input mode,
when using postmeme.js, or postmsg.js with the -P option.

Not done: adding or parsing the "Format" parameter in MIME content-type
message header fields per RFC-3676.