SlyEdit: Better behavior for editing general files, to support editing SSH keys & such. Support for appending a meme to a message by typing /m on an empty line and pressing enter.
Even though text is wrapped to fit the user's terminal, when editing something, SlyEdit
saves new text on a single line rather than saving the wrapped text, so that readers can
wrap text to fit the user's terminal. Normally, when saving a message for a messagebase
(such as email or a sub-board, SlyEdit adds a space between each line, which makes sense
due to SlyEdit's line wrapping, which wraps whole words to the next line.  Now, if there
is no internal messagebase code, SlyEdit knows it's simply editing a file and not posting
in a messagebase and does not add a space between wrapped lines, ensuring that the text
all goes together as typed in SlyEdit.  Text displayed in SlyEdit will still be wrapped
to the user's terminal width, but it should be saved as continuous text (except, of
course, with blank lines in between).  The option to not wrap to the user's terminal and
scroll to the right instead may be something added to SlyEdit in the future.

Other changes & new features:

Color/attribute codes are no longer removed from quote lines. Although quote lines still
appear in SlyEdit with one (configurable) color, the quoted text will retain color/attribute
codes when the message is posted.

New feature: A 'meme' can be added to the message by typing /m on an empty line by itself
and pressing enter. This is basically the behavior of postmeme.js, except SlyEdit adds the
meme to the message at the current cursor/line location. New meme configuration options
have been added to SlyEdit.cfg.

Bug fix: When closing the User Settings dialog, quote lines are refreshed with the
configured quote line color (instead of with any color codes included in those lines)

Bug fix: When closing the User Settings dialog, existing message lines are refreshed better;
sometimes, there were small parts of the beginnings of some lines that were blanked out.

Bug fix (sort of): DCT mode File menu wasn't drawing when the user presses the ESC key to
bring up the menu. I don't know why that's happening. I found a kludge that seems to get
the File menu to display again when pressing ESC.

Fix/better behavior: On startup, SlyEdit now sets the 'normal' attribute on the console to
clear away any background color or other attribute(s) that may have been set.