Add "maximum message age" in days option for AreaMgr %RESCAN use
Support for -D=<days> in message subject and D=<days> option in
%RESCAN and
%RESCAN <area-tag> commands

The -R subject option can be used in combination with the %RESCAN D= option
(or vice versa) to both limit the number *and* the age of messages, if desired.

Include both the R=<cnt> and D=<days> option in the %RESCAN command is not
supported (but could be if deemed important).

Resolves issue #929