Include missing header fields (from msglist.js source-view usage)
We were missing a lot of header field output that's otherwise available
when viewing message headers via "smbutil v" or smb_msghdr_str_list()
(the method used to dump header fields via the built-in operator 'H'eader
view command).

If there's a sender or recipient extension, include that (as a comment)
in the From and To header field output (e.g. "From: Sysop (#1)").

Include the recipient network address in the To line (as a comment when
it's a FidoNet or QWKnet address).

When the sender network address type is FidoNet or QWKnet (i.e. doesn't
contain an '@'), include it as a comment rather than "<address>".

Include the X-FTN-BBSID header field, if it exists

Include any SMB FIDOCTRL header fields (e.g. "Via" kludges) as "X-FTN-Kludge"
header fields in the output.