Work around bug in issue #936
The regexp engine used in aarch64 builds has a bug where /x[^x]*?$/
will not match when /x[^x]*$/ will, which causes issues with lockfiles.

Simply removing the ? (which in this case should not change the
results) resolves it.

Other cases of *?$ in regexps should be investigated to ensure this
is the only workaround (this is the only case of an inverted alternative
followed by *?$ I can find though).