Fix issue #942
Session is initialized to HTTP/0.9 (which doesn't send headers) and
the HEAD method (which doesn't send a body). This was caught in
writebuf(), which returned 0 due to the body not being sent, which
caused the error to be logged.

Set the values to HTTP/1.0 and GET so we can call send_error()
before we read a request line and actually send an error.

Additionally, add a LOG_INFO to writebuf() when it returns 0 due
to body being disabled to make this kind of thing easier to find
in the future.

Finally, don't attempt to send an error body if send_headers()
fails or if we should not send a body.

This would also allow us to send a 408 if we wanted to, but I've
never been convinced that's useful.