Support a filter (*.can) file 'quiet' attribute (e.g. "q=1")
A dubious feature. This will squelch any "CLIENT BLOCKED" type log messages
assocated with a filtered client. Importantly, other log messages related to
the client connection won't be (can't be) squelched, so this isn't going to
really be a useful feature in my opinion, but when sysops ask for this (e.g.
silently-blocking specific hostnames), I want to be able to demonstrate why
that's not really practical.

The syntax is "<filter-string><tab>q=1" if you want to play with it.