Better error reporting/handling when the ssl.cert file doesn't exist
Since "Create self-signed certificate" defaults to No, it's normal/expected to
not have an ssl.cert file in many cases and the logged errors were not very
helpful. We should probably change the default for this setting to Yes
(and have letsyncrypt.js disable it) or when auto-disable any SSH/TLS
functionality that will expect a certificate file when the file doens't exist
during initialization (startup or recycle). At least this change is a little
bit of an improvement (help to the unknowing sysop). No longer just logging
this unhelpful spew:
7/8 21:23:05 mail 0055 SEND connecting to port 25 on mail.synchro.net [71.95.196.36]
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:06 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:07 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:08 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:10 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:14 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:22 mail Failed to open/read TLS certificate: /sbbs/ctrl/ssl.cert
7/8 21:23:22 mail 0055 SEND/TLS [mail.synchro.net] ERROR 'Data has not been initialised' (-11) setting private key

Thanks to Sam Alexander (BLUEBOX) for pointing this bad behavior out.