Set cfg_changed flag (so msgbases are created/updated) when changing int-codes
When changing message or file area internal code prefixes or suffixes, the
message base files need to be (re)created (e.g. in data/subs or data/dirs).
The 'cfg_changed' flag is the thing that indicates that we need to do that for
each message or file area and we weren't setting that flag in these cases.

Note: this doesn't rename the message base files, so if they have content
(messages or files), they'll just remain in the old internal code named files
and require the sysop to deal with manually.

Reported in IRC by finglonger.