Update to previous commit (which didn't fix reported issue)
The trailing slash on the sub-directory name was still part of the string
being compared (e.g. to internal codes), so that wasn't going to fix it.
Kind of refactored this since it was removing and re-adding and then removing
the trailing the slash again. And we didn't need to write the dir.code here
since that's written in the loop that follows.

This should fix the issue that finglonger reported on IRC.